restructure_project
create packages: model, control, utility
