+++ FunPlus/app/src/main/AndroidManifest.xml
@@ -24,9 +24,9 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".PrizeListActivity"></activity>
-        <activity android:name=".ArActivity" />
-        <activity android:name=".MainActivity">
+        <activity android:name=".control.PrizeListActivity"></activity>
+        <activity android:name=".control.ArActivity" />
+        <activity android:name=".control.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
