+++ FunPlus/app/src/main/java/com/example/funplus/control/LetterFrag.kt
@@ -1,12 +1,11 @@
-package com.example.funplus
+package com.example.funplus.control
 
-import android.content.Context
-import android.net.Uri
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.example.funplus.R
 
 class LetterFrag : Fragment(){
