+++ FunPlus/app/src/main/java/com/example/funplus/control/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.funplus
+package com.example.funplus.control
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
@@ -6,6 +6,7 @@ import android.util.Log
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import androidx.fragment.app.FragmentTransaction
+import com.example.funplus.R
 import kotlinx.android.synthetic.main.activity_main.*
 
 const val TAG = "DBG"
