+++ FunPlus/app/src/main/java/com/example/funplus/model/Game.kt
@@ -1,4 +1,4 @@
-package com.example.funplus
+package com.example.funplus.model
 
 data class Game(val num1: Int, val num2: Int, val sign: String){
     override fun toString(): String {
