+++ FunPlus/app/src/main/java/com/example/funplus/model/GameDB.kt
@@ -1,4 +1,4 @@
-package com.example.funplus
+package com.example.funplus.model
 
 import android.content.Context
 import androidx.room.Database
